Removed remote-scripts
Motivation for removing: to complex to maintain, ugly messages in logs
because of `export PATH ...`

To add new script: use role files, as example, see how lspartitions.sh
was converted to lincommon-partitions.yaml

To add new binary: no w/a for now. Later can be implemented via
resources.

Disadvantages: won't work in case if /tmp is mounted with noexec