hwswa2.server.req._generate_join_logical() now discards non-unique requirements
if AND or OR requirement consists of logically the same requirements, it
discards them. If only one requirement is to be included into AND/OR,
result will be this requirement.

F.e., if you have requirement for the OS to be CentOS 6.7 in two roles,
and one server has both roles, _generate_join_logical() will try to
construct AND requirement from those two, will discard one and return
one OS requirement instead of AND with two OS in it.